+++ src/context.rs
@@ -1,19 +1,20 @@
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
 use std::{
+    collections::HashSet,
     ffi::OsString,
     io::{self, Write as _},
     path::PathBuf,
 };
 
-use anyhow::{Result, bail};
+use anyhow::{Context as _, Result, bail};
 use camino::Utf8PathBuf;
 
 use crate::{
     cargo::Workspace,
     cli::{self, Args, Subcommand},
     env,
-    metadata::{Metadata, PackageId},
+    metadata::{Package, PackageId},
     process::ProcessBuilder,
     regex_vec::{RegexVec, RegexVecBuilder},
     term,
@@ -217,12 +218,8 @@ impl Context {
             }
         };
 
-        let workspace_members = WorkspaceMembers::new(
-            &args.exclude,
-            &args.exclude_from_report,
-            &args.package,
-            &ws.metadata,
-        );
+        let workspace_members =
+            WorkspaceMembers::new(&ws, &args.exclude_from_report, &args.package, args.workspace)?;
         if workspace_members.included.is_empty() {
             bail!("no crates to be measured for coverage");
         }
@@ -298,33 +295,246 @@ pub(crate) struct WorkspaceMembers {
 
 impl WorkspaceMembers {
     fn new(
-        exclude: &[String],
+        ws: &Workspace,
         exclude_from_report: &[String],
         package: &[String],
-        metadata: &Metadata,
-    ) -> Self {
+        workspace: bool,
+    ) -> Result<Self> {
         let mut excluded = vec![];
         let mut included = vec![];
-        let has_exclude = !exclude.is_empty() || !exclude_from_report.is_empty();
-        for &id in &metadata.workspace_members {
-            let pkg = &metadata[id];
-            // --exclude flag doesn't handle `name:version` format
-            if has_exclude
-                && (exclude.contains(&pkg.name) || exclude_from_report.contains(&pkg.name))
-            {
-                excluded.push(id);
-            } else if package.is_empty()
-                || package.contains(&pkg.name)
-                || package.contains(&format!("{}:{}", &pkg.name, &pkg.version))
-            {
-                included.push(id);
-            } else {
-                excluded.push(id);
+        // Refs: https://github.com/rust-lang/cargo/blob/0d08b955e5f6171f81e5268b91a7d70f2e94b62f/src/cargo/ops/cargo_compile/packages.rs
+        let mut opt_out = if exclude_from_report.is_empty() {
+            None
+        } else {
+            Some(find_ids(ws, exclude_from_report)?)
+        };
+        if opt_out.is_none() && workspace {
+            included.extend_from_slice(&ws.metadata.workspace_members);
+        } else {
+            let mut opt_in = if package.is_empty() { None } else { Some(find_ids(ws, package)?) };
+            'outer: for &id in &ws.metadata.workspace_members {
+                if let Some((ids, pats)) = &mut opt_out {
+                    // --exclude
+                    if ids.contains(&id) {
+                        excluded.push(id);
+                        continue;
+                    }
+                    let name = &ws.metadata[id].name;
+                    for pat in pats {
+                        if pat.0.matches(name) {
+                            excluded.push(id);
+                            pat.1 = true;
+                            continue 'outer;
+                        }
+                    }
+                }
+                if workspace {
+                    // --workspace
+                    included.push(id);
+                } else if let Some((ids, pats)) = &mut opt_in {
+                    // --package
+                    if ids.contains(&id) {
+                        included.push(id);
+                        continue;
+                    }
+                    let name = &ws.metadata[id].name;
+                    for pat in pats {
+                        if pat.0.matches(name) {
+                            included.push(id);
+                            pat.1 = true;
+                            continue 'outer;
+                        }
+                    }
+                    excluded.push(id);
+                } else if let Some(current_package) = ws.current_package {
+                    // root of non-virtual workspace or member of virtual workspace
+                    if id == current_package {
+                        included.push(id);
+                    } else {
+                        excluded.push(id);
+                    }
+                } else {
+                    // root of virtual workspace
+                    included.push(id);
+                }
+            }
+
+            if let Some((_, pats)) = &opt_out {
+                for (pat, matched) in pats {
+                    if !matched {
+                        warn!("not found package pattern '{pat}' in workspace");
+                    }
+                }
+            }
+            if let Some((_, pats)) = &opt_in {
+                for (pat, matched) in pats {
+                    if !matched {
+                        warn!("not found package pattern '{pat}' in workspace");
+                    }
+                }
             }
         }
 
-        Self { excluded, included }
+        Ok(Self { excluded, included })
     }
+}
+
+fn find_ids(
+    ws: &Workspace,
+    list: &[String],
+) -> Result<(HashSet<PackageId>, Vec<(glob::Pattern, bool)>)> {
+    let mut ids = HashSet::with_capacity(list.len());
+    let mut patterns = vec![];
+    for e in list {
+        let mut found = false;
+        for &id in &ws.metadata.workspace_members {
+            if match_pkg_spec(&ws.metadata[id], e)? {
+                ids.insert(id);
+                found = true;
+                break;
+            }
+        }
+        if !found {
+            if e.contains(['*', '?', '[', ']']) {
+                patterns.push((
+                    glob::Pattern::new(e)
+                        .with_context(|| format!("cannot build glob pattern from `{e}`"))?,
+                    false,
+                ));
+            } else {
+                warn!("not found package '{e}' in workspace");
+            }
+        }
+    }
+    Ok((ids, patterns))
+}
+
+fn match_pkg_spec(pkg: &Package, name_or_spec: &str) -> Result<bool> {
+    /*
+    Refs: https://doc.rust-lang.org/1.93.0/cargo/reference/pkgid-spec.html
+        spec := pkgname |
+            [ kind "+" ] proto "://" hostname-and-path [ "?" query] [ "#" ( pkgname | semver ) ]
+        query = ( "branch" | "tag" | "rev" ) "=" ref
+        pkgname := name [ ("@" | ":" ) semver ]
+        semver := digits [ "." digits [ "." digits [ "-" prerelease ] [ "+" build ]]]
+
+        kind = "registry" | "git" | "path"
+        proto := "http" | "git" | "file" | ...
+    */
+    fn split_spec(s: &str) -> Option<(&str, &str, Option<&str>, Option<&str>)> {
+        let (proto_etc, hostname_and_path_etc) = s.split_once("://")?;
+        let proto = proto_etc.split_once('+').unwrap_or(("", proto_etc)).1; // drop kind
+        let (hostname_and_path_etc, pkgname_or_semver) =
+            hostname_and_path_etc.split_once('#').unwrap_or((hostname_and_path_etc, ""));
+        let (hostname_and_path, query) =
+            hostname_and_path_etc.split_once('?').unwrap_or((hostname_and_path_etc, ""));
+        Some((
+            proto,
+            hostname_and_path,
+            if query.is_empty() { None } else { Some(query) },
+            if pkgname_or_semver.is_empty() { None } else { Some(pkgname_or_semver) },
+        ))
+    }
+    fn split_semver(
+        s: &str,
+    ) -> Option<(&str, Option<(&str, Option<(&str, Option<&str>, Option<&str>)>)>)> {
+        let mut digits = s.splitn(3, '.');
+        let major = digits.next()?;
+        let Some(minor) = digits.next() else {
+            return Some((major, None));
+        };
+        let Some(patch_etc) = digits.next() else {
+            return Some((major, Some((minor, None))));
+        };
+        let (patch_etc, meta) = patch_etc.split_once('+').unwrap_or((patch_etc, ""));
+        let (patch, pre) = patch_etc.split_once('-').unwrap_or((patch_etc, ""));
+        Some((
+            major,
+            Some((
+                minor,
+                Some((
+                    patch,
+                    if pre.is_empty() { None } else { Some(pre) },
+                    if meta.is_empty() { None } else { Some(meta) },
+                )),
+            )),
+        ))
+    }
+    let name = &pkg.name;
+    let p = name_or_spec;
+    let (version, full_version) = if p.starts_with(name) {
+        if p.len() == name.len() {
+            return Ok(true); // version omitted
+        }
+        if !matches!(p.as_bytes().get(name.len()), Some(&b'@' | &b':')) {
+            return Ok(false); // pkgname unmatched
+        }
+        (&p[name.len() + 1..], &*pkg.version)
+    } else {
+        let p = p.trim_ascii_end(); // pkgid may contains trailing newline (e.g., when pkgid is got from `cargo pkgid -p <package>`)
+        let full = &*pkg.id;
+        let Some((proto, hostname_and_path, query, pkgname_or_semver)) = split_spec(p) else {
+            return Ok(false); // p is not pkg spec
+        };
+        let Some((full_proto, full_hostname_and_path, full_query, full_pkgname_or_semver)) =
+            split_spec(full)
+        else {
+            bail!("invalid pkg spec ({full}) from cargo-metadata")
+        };
+        if proto != full_proto || hostname_and_path != full_hostname_and_path {
+            return Ok(false); // proto or hostname-and-path unmatched
+        }
+        if query.is_some() && query != full_query {
+            return Ok(false); // query unmatched
+        }
+        let Some(pkgname_or_semver) = pkgname_or_semver else {
+            return Ok(true); // pkgname | semver omitted
+        };
+        let Some(full_pkgname_or_semver) = full_pkgname_or_semver else {
+            return Ok(false); // extra pkgname | semver
+        };
+        match (
+            pkgname_or_semver.split_once(['@', ':']),
+            full_pkgname_or_semver.split_once(['@', ':']),
+        ) {
+            (Some((pkgname, semver)), Some((full_pkgname, full_semver))) => {
+                if pkgname != full_pkgname {
+                    return Ok(false); // pkgname unmatched
+                }
+                (semver, full_semver)
+            }
+            (Some(_), None) => return Ok(false), // extra semver
+            (None, _) => return Ok(true),        // pkgname omitted or no pkgname in spec
+        }
+    };
+    let Some((major, minor_etc)) = split_semver(version) else {
+        warn!("invalid pkg version ({version}) from --package");
+        return Ok(false); // invalid version
+    };
+    let Some((full_major, Some((full_minor, Some((full_patch, full_pre, full_meta)))))) =
+        split_semver(full_version)
+    else {
+        bail!("invalid pkg version ({full_version}) from cargo-metadata")
+    };
+    if major != full_major {
+        return Ok(false); // major unmatched
+    }
+    let Some((minor, patch_etc)) = minor_etc else {
+        return Ok(true); // minor version omitted
+    };
+    if minor != full_minor {
+        return Ok(false); // minor unmatched
+    }
+    let Some((patch, pre, meta)) = patch_etc else {
+        return Ok(true); // patch version omitted
+    };
+    if patch != full_patch
+        || pre.is_some() && pre != full_pre
+        || meta.is_some() && meta != full_meta
+    {
+        return Ok(false); // patch or pre or meta unmatched
+    }
+    Ok(true)
 }
 
 // Adapted from https://github.com/rust-lang/miri/blob/dba35d2be72f4b78343d1a0f0b4737306f310672/cargo-miri/src/util.rs#L181-L204
@@ -350,4 +560,102 @@ fn ask_to_run(cmd: &ProcessBuilder, ask: bool, text: &str) -> Result<()> {
 
     cmd.run()?;
     Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::{Package, match_pkg_spec};
+
+    #[test]
+    fn test_match_pkg_spec() {
+        // Examples are from https://doc.rust-lang.org/1.93.0/cargo/reference/pkgid-spec.html#example-specifications
+
+        // crates.io
+        let pkg = &Package {
+            id: "registry+https://github.com/rust-lang/crates.io-index#regex@1.4.3".to_owned(),
+            name: "regex".to_owned(),
+            version: "1.4.3".to_owned(),
+            targets: vec![],
+            manifest_path: "".into(),
+        };
+        // name
+        assert!(match_pkg_spec(pkg, "regex").unwrap());
+        assert!(!match_pkg_spec(pkg, "regex-syntax").unwrap());
+        // name+version
+        assert!(match_pkg_spec(pkg, "regex@1").unwrap());
+        assert!(match_pkg_spec(pkg, "regex@1.4").unwrap());
+        assert!(match_pkg_spec(pkg, "regex@1.4.3").unwrap());
+        assert!(match_pkg_spec(pkg, "regex:1.4").unwrap());
+        assert!(!match_pkg_spec(pkg, "regex@2").unwrap());
+        assert!(!match_pkg_spec(pkg, "regex@1.5").unwrap());
+        assert!(!match_pkg_spec(pkg, "regex@1.4.2").unwrap());
+        assert!(!match_pkg_spec(pkg, "regex@1.4.4").unwrap());
+        // spec
+        assert!(match_pkg_spec(pkg, "https://github.com/rust-lang/crates.io-index#regex").unwrap());
+        assert!(
+            match_pkg_spec(pkg, "https://github.com/rust-lang/crates.io-index#regex@1.4.3")
+                .unwrap()
+        );
+        assert!(
+            match_pkg_spec(pkg, "https://github.com/rust-lang/crates.io-index#regex@1.4").unwrap()
+        );
+        assert!(
+            match_pkg_spec(
+                pkg,
+                "registry+https://github.com/rust-lang/crates.io-index#regex@1.4.3"
+            )
+            .unwrap()
+        );
+
+        // git
+        let pkg = &Package {
+            id: "git+ssh://git@github.com/rust-lang/regex.git?branch=dev#regex@1.4.3".to_owned(),
+            name: "regex".to_owned(),
+            version: "1.4.3".to_owned(),
+            targets: vec![],
+            manifest_path: "".into(),
+        };
+        assert!(match_pkg_spec(pkg, "regex").unwrap());
+        assert!(
+            match_pkg_spec(pkg, "ssh://git@github.com/rust-lang/regex.git#regex@1.4.3").unwrap()
+        );
+        assert!(
+            match_pkg_spec(pkg, "git+ssh://git@github.com/rust-lang/regex.git#regex@1.4.3")
+                .unwrap()
+        );
+        assert!(
+            match_pkg_spec(
+                pkg,
+                "git+ssh://git@github.com/rust-lang/regex.git?branch=dev#regex@1.4.3"
+            )
+            .unwrap()
+        );
+        let pkg = &Package {
+            id: "git+https://github.com/rust-lang/cargo#0.52.0".to_owned(),
+            name: "cargo".to_owned(),
+            version: "0.52.0".to_owned(),
+            targets: vec![],
+            manifest_path: "".into(),
+        };
+        assert!(match_pkg_spec(pkg, "https://github.com/rust-lang/cargo#0.52.0").unwrap());
+        assert!(match_pkg_spec(pkg, "git+https://github.com/rust-lang/cargo#0.52.0").unwrap());
+        assert!(
+            !match_pkg_spec(pkg, "https://github.com/rust-lang/cargo#cargo-platform@0.1.2")
+                .unwrap()
+        );
+
+        // local
+        let pkg = &Package {
+            id: "path+file:///path/to/my/project/foo#1.1.8".to_owned(),
+            name: "foo".to_owned(),
+            version: "1.1.8".to_owned(),
+            targets: vec![],
+            manifest_path: "".into(),
+        };
+        assert!(match_pkg_spec(pkg, "foo").unwrap());
+        assert!(match_pkg_spec(pkg, "file:///path/to/my/project/foo").unwrap());
+        assert!(match_pkg_spec(pkg, "file:///path/to/my/project/foo#1.1.8").unwrap());
+        assert!(match_pkg_spec(pkg, "path+file:///path/to/my/project/foo#1.1").unwrap());
+        assert!(match_pkg_spec(pkg, "path+file:///path/to/my/project/foo#1.1.8").unwrap());
+    }
 }